cmd/crawler/main_test.go
package main

import (
	"os"
	"testing"

	"github.com/gocolly/colly"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
	"go.uber.org/zap"
)

// MockDrug is a mocked object for termmatcher.Related
type MockDrug struct {
	mock.Mock
}

// Related is a mocked implementation for termmatcher.Related
func (m *MockDrug) Related(href string) bool {
	args := m.Called(href)
	return args.Bool(0)
}

// MockRedisWrapper is a mocked object for rediswrapper functions
type MockRedisWrapper struct {
	mock.Mock
}

// SAdd is a mocked implementation for rediswrapper.SAdd
func (m *MockRedisWrapper) SAdd(href string) error {
	args := m.Called(href)
	return args.Error(0)
}

// PublishHref is a mocked implementation for rediswrapper.PublishHref
func (m *MockRedisWrapper) PublishHref(stream, href, group string) error {
	args := m.Called(stream, href, group)
	return args.Error(0)
}

// Del is a mocked implementation for rediswrapper.Del
func (m *MockRedisWrapper) Del() error {
	args := m.Called()
	return args.Error(0)
}

func TestParseCommandLineArguments(t *testing.T) {
	// Test case 1: All required arguments provided
	args1 := []string{"./crawler", "-url=https://www.example.com", "-search=search-term-1,search-term-2"}
	os.Args = args1
	config1, err1 := parseCommandLineArguments()
	assert.Equal(t, "https://www.example.com", config1.URL, "Expected crawlURL to match")
	assert.Equal(t, "search-term-1,search-term-2", config1.SearchTerms, "Expected search terms to match")
	assert.NoError(t, err1, "Expected no error")

	// ...
}

func TestCreateLogger(t *testing.T) {
	logger := createLogger()

	// Check if the logger is not nil
	assert.NotNil(t, logger, "Expected logger not to be nil")

	// Log a message
	logger.Info("Test log message")

	// No need to check for a return value from logger.Info
}

func TestCreateRedisClient(t *testing.T) {
	// Set up environment variables for testing
	os.Setenv("REDIS_HOST", "localhost")
	os.Setenv("REDIS_PORT", "6379")
	os.Setenv("REDIS_AUTH", "")

	// Clean up environment variables after the test
	defer func() {
		os.Unsetenv("REDIS_HOST")
		os.Unsetenv("REDIS_PORT")
		os.Unsetenv("REDIS_AUTH")
	}()

	// Create the Redis client
	redisClient := createRedisClient()

	// Assertions
	assert.NotNil(t, redisClient, "Expected redisClient not to be nil")
	defer redisClient.Close()

	// Additional assertions based on your Redis client configuration
	// For example, you can check if the client's options match your expectations
	assert.Equal(t, "localhost:6379", redisClient.Options().Addr, "Expected Redis address to match")
	assert.Equal(t, "", redisClient.Options().Password, "Expected empty Redis password")
	// Add more assertions as needed
}

func TestConfigureCollector(t *testing.T) {
	domain := "example.com"                                   // Replace with your desired domain
	logger := createLogger()                                  // Create a logger for testing
	collector := configureCollector([]string{domain}, logger) // Pass the domain and logger

	// Assertions
	assert.NotNil(t, collector, "Expected collector not to be nil")
	assert.True(t, collector.Async, "Expected collector to be asynchronous")
	assert.Equal(t, 2, collector.MaxDepth, "Expected MaxDepth to be 2")
	// You can add more assertions based on your requirements
}

func TestSetupCrawlingLogic(t *testing.T) {
	// Create a new collector
	collector := colly.NewCollector()

	// Create a mock logger
	logger := &zap.SugaredLogger{}

	// Define search terms for testing
	searchTerms := []string{
		"DRUG",
		"SMOKE JOINT",
		"GROW OP",
		"CANNABI",
		"IMPAIR",
		"SHOOT",
		"FIREARM",
		"MURDER",
		"COCAIN",
		"POSSESS",
		"BREAK ENTER",
	}

	// Inject the mocked instances and search terms into your setupCrawlingLogic function
	setupCrawlingLogic(collector, logger, searchTerms)

	// Your test assertions here
	// ...TestSetupCrawlingLogic

	// Assert that the expectations of your mocks are met
	// ...

	// Clean up any resources
	// ...
}

func TestMain(m *testing.M) {
	// Run tests and exit with the result
	os.Exit(m.Run())
}
